Changes made to most items in model-training folder to tailor files for our project
