+++ model-training/cli.py
@@ -6,7 +6,7 @@ import google.cloud.aiplatform as aip
 GCP_PROJECT = os.environ["GCP_PROJECT"]
 GCP_REGION = os.environ["GCP_REGION"]
 GCS_BUCKET_URI = os.environ["GCS_BUCKET_URI"]
-
+WANDB_KEY = os.environ["WANDB_KEY"]
 
 def generate_uuid(length: int = 8) -> str:
     return "".join(random.choices(string.ascii_lowercase + string.digits, k=length))
@@ -16,25 +16,25 @@ def generate_uuid(length: int = 8) -> str:
 aip.init(project=GCP_PROJECT, location=GCP_REGION, staging_bucket=GCS_BUCKET_URI)
 
 job_id = generate_uuid()
-DISPLAY_NAME = "mushroom_" + job_id
+DISPLAY_NAME = "model_" + job_id
 
 TRAIN_IMAGE = "us-docker.pkg.dev/vertex-ai/training/tf-gpu.2-12.py310:latest"
 
 job = aip.CustomPythonPackageTrainingJob(
     display_name=DISPLAY_NAME,
-    python_package_gcs_uri=f"{GCS_BUCKET_URI}/mushroom-app-trainer.tar.gz",
+    python_package_gcs_uri=f"{GCS_BUCKET_URI}/trainer_package/snapnutrition-trainer.tar.gz",
     python_module_name="trainer.task",
     container_uri=TRAIN_IMAGE,
     project=GCP_PROJECT,
 )
 
-CMDARGS = ["--epochs=15", "--batch_size=16"]
+CMDARGS = ["--config=model_config.yml", f"--wandb_key={WANDB_KEY}"]
 MODEL_DIR = GCS_BUCKET_URI
 TRAIN_COMPUTE = "n1-standard-4"
 TRAIN_GPU = "NVIDIA_TESLA_T4"
 TRAIN_NGPU = 1
 
-print(f"{GCS_BUCKET_URI}/mushroom-app-trainer.tar.gz")
+print(f"{GCS_BUCKET_URI}/trainer_package/snapnutrition-trainer.tar.gz")
 print(TRAIN_IMAGE)
 
 # Run the training job on Vertex AI
