+++ model-training/package-trainer.sh
@@ -1,4 +1,4 @@
 rm -f trainer.tar trainer.tar.gz
 tar cvf trainer.tar package
 gzip trainer.tar
-gsutil cp trainer.tar.gz $GCS_BUCKET_URI/snap-nutrition-model-trainer.tar.gz
+gsutil cp trainer.tar.gz $GCS_BUCKET_URI/trainer_package/snapnutrition-trainer.tar.gz
