+++ model-training/package/PKG-INFO
@@ -1,6 +1,6 @@
 Metadata-Version: 1.0
 
-Name: Snap-Nutrition App Trainer
+Name: SnapNutrition App Trainer
 
 Version: 0.0.1
