+++ model-training/package/setup.py
@@ -2,10 +2,14 @@ from setuptools import find_packages
 from setuptools import setup
 
 
-REQUIRED_PACKAGES = ["wandb==0.15.11", "tensorflow-hub==0.14.0", "google-cloud-storage", "pandas"]
+REQUIRED_PACKAGES = [
+    "wandb==0.15.11", 
+    "google-cloud-storage",
+    "pyyaml" 
+]
 
 setup(
-    name="snap-nutrition-model-trainer",
+    name="snapnutrition-model-trainer",
     version="0.0.1",
     install_requires=REQUIRED_PACKAGES,
     packages=find_packages(),
