+++ model-training/package/trainer/task.py
@@ -1,382 +1,295 @@
-import argparse
 import os
-import requests
-import zipfile
-import tarfile
 import time
+import math
+import uuid
+import argparse
+import yaml
+from google.cloud import storage
 
 # Tensorflow
+os.environ['TF_CPP_MIN_LOG_LEVEL'] = '3' # silence tf info, error, warning messages
 import tensorflow as tf
-from tensorflow import keras
-from tensorflow.keras.models import Model, Sequential
-from tensorflow.keras.utils import to_categorical
-from tensorflow.python.keras import backend as K
-from tensorflow.python.keras.utils.layer_utils import count_params
-
-# sklearn
-from sklearn.model_selection import train_test_split
-
-# Tensorflow Hub
-import tensorflow_hub as hub
-
-# W&B
-import wandb
-from wandb.keras import WandbCallback, WandbMetricsLogger
-
-
-# Setup the arguments for the trainer task
-parser = argparse.ArgumentParser()
-parser.add_argument(
-    "--model-dir", dest="model_dir", default="test", type=str, help="Model dir."
-)
-parser.add_argument("--lr", dest="lr", default=0.001, type=float, help="Learning rate.")
-parser.add_argument(
-    "--model_name",
-    dest="model_name",
-    default="mobilenetv2",
-    type=str,
-    help="Model name",
-)
-parser.add_argument(
-    "--train_base",
-    dest="train_base",
-    default=False,
-    action="store_true",
-    help="Train base or not",
-)
-parser.add_argument(
-    "--epochs", dest="epochs", default=10, type=int, help="Number of epochs."
-)
-parser.add_argument(
-    "--batch_size", dest="batch_size", default=16, type=int, help="Size of a batch."
-)
-parser.add_argument(
-    "--wandb_key", dest="wandb_key", default="16", type=str, help="WandB API Key"
-)
-args = parser.parse_args()
-
-# TF Version
-print("tensorflow version", tf.__version__)
+from tensorflow.keras.models import Sequential
+from tensorflow.keras.layers import Conv2D, Dense, Dropout 
+from tensorflow.keras.layers import Flatten, MaxPooling2D
+from tensorflow.keras.optimizers import Adam
+print(f'Tensorflow Version: {tf.__version__}')
 print("Eager Execution Enabled:", tf.executing_eagerly())
+
 # Get the number of replicas
 strategy = tf.distribute.MirroredStrategy()
-print("Number of replicas:", strategy.num_replicas_in_sync)
+print("Number of Replicas:", strategy.num_replicas_in_sync)
 
 devices = tf.config.experimental.get_visible_devices()
 print("Devices:", devices)
-print(tf.config.experimental.list_logical_devices("GPU"))
+print(tf.config.experimental.list_logical_devices('GPU'))
 
-print("GPU Available: ", tf.config.list_physical_devices("GPU"))
+print("GPU Available: ", tf.config.list_physical_devices('GPU'))
 print("All Physical Devices", tf.config.list_physical_devices())
 
-
-# Utils functions
-def download_file(packet_url, base_path="", extract=False, headers=None):
-    if base_path != "":
-        if not os.path.exists(base_path):
-            os.mkdir(base_path)
-    packet_file = os.path.basename(packet_url)
-    with requests.get(packet_url, stream=True, headers=headers) as r:
-        r.raise_for_status()
-        with open(os.path.join(base_path, packet_file), "wb") as f:
-            for chunk in r.iter_content(chunk_size=8192):
-                f.write(chunk)
-
-    if extract:
-        if packet_file.endswith(".zip"):
-            with zipfile.ZipFile(os.path.join(base_path, packet_file)) as zfile:
-                zfile.extractall(base_path)
-        else:
-            packet_name = packet_file.split(".")[0]
-            with tarfile.open(os.path.join(base_path, packet_file)) as tfile:
-                tfile.extractall(base_path)
+# W&B
+import wandb
+from wandb.keras import WandbCallback
 
 
-# Download Data
-start_time = time.time()
-download_file(
-    "https://github.com/dlops-io/datasets/releases/download/v1.0/mushrooms_3_labels.zip",
-    base_path="datasets",
-    extract=True,
-)
-execution_time = (time.time() - start_time) / 60.0
-print("Download execution time (mins)", execution_time)
-
-# Load Data
-base_path = os.path.join("datasets", "mushrooms")
-label_names = os.listdir(base_path)
-print("Labels:", label_names)
-
-# Number of unique labels
-num_classes = len(label_names)
-# Create label index for easy lookup
-label2index = dict((name, index) for index, name in enumerate(label_names))
-index2label = dict((index, name) for index, name in enumerate(label_names))
-
-# Generate a list of labels and path to images
-data_list = []
-for label in label_names:
-    # Images
-    image_files = os.listdir(os.path.join(base_path, label))
-    data_list.extend([(label, os.path.join(base_path, label, f)) for f in image_files])
-
-print("Full size of the dataset:", len(data_list))
-print("data_list:", data_list[:5])
-
-# Load X & Y
-# Build data x, y
-data_x = [itm[1] for itm in data_list]
-data_y = [itm[0] for itm in data_list]
-print("data_x:", len(data_x))
-print("data_y:", len(data_y))
-print("data_x:", data_x[:5])
-print("data_y:", data_y[:5])
-
-# Split Data
-test_percent = 0.10
-validation_percent = 0.2
-
-# Split data into train / test
-train_validate_x, test_x, train_validate_y, test_y = train_test_split(
-    data_x, data_y, test_size=test_percent
+# Parse command line arguments
+parser = argparse.ArgumentParser(
+    description='Build a VGG-like CNN model from a YAML configuration file.'
 )
 
-# Split data into train / validate
-train_x, validate_x, train_y, validate_y = train_test_split(
-    train_validate_x, train_validate_y, test_size=test_percent
+parser.add_argument(
+    '-c',
+    '--config', 
+    required=True,
+    type=str, 
+    help='Path to the YAML configuration file'
 )
 
-print("train_x count:", len(train_x))
-print("validate_x count:", len(validate_x))
-print("test_x count:", len(test_x))
+parser.add_argument(
+    '-w',
+    '--wandb_key',
+    required=True,
+    type=str, 
+    help="Weights and Biases API Key"
+)
 
-# Login into wandb
+args = parser.parse_args()
+
+
+# Define globals
+GCS_BUCKET_NAME = os.path.basename(os.environ["GCS_BUCKET_URI"])
+GCP_PROJECT = os.environ["GCP_PROJECT"]
+tfrecords_folder = "data/tf_records/180_by_180/"
+models_folder = "models/"
+
+
+# Login to Weights and Biases
 wandb.login(key=args.wandb_key)
 
 
-# Create TF Datasets
-def get_dataset(image_width=224, image_height=224, num_channels=3, batch_size=32):
-    # Load Image
-    def load_image(path, label):
-        image = tf.io.read_file(path)
-        image = tf.image.decode_jpeg(image, channels=num_channels)
-        image = tf.image.resize(image, [image_height, image_width])
-        return image, label
+def build_model(
+    input_shape,
+    n_classes, 
+    n_filters=16,
+    dense_nodes=1024, 
+    conv_dropout=0.25, 
+    dense_dropout=0.5,  
+    filter_size=(3,3), 
+    pool_size=(2,2), 
+    output_activation='softmax',
+    model_name='model' 
+    ):
 
-    # Normalize pixels
-    def normalize(image, label):
-        image = image / 255
-        return image, label
+    """
+    This function builds a simplified vgg-like convolutional model
+   
+    """
+    model = Sequential(
+        [
+            Conv2D(
+                n_filters, filter_size, padding="same", activation="relu",
+                input_shape=input_shape,
+            ),
+            Conv2D(n_filters, filter_size, padding="same", activation="relu"),
+            MaxPooling2D(pool_size=pool_size),
+            Dropout(conv_dropout),
+            
+            Conv2D(n_filters*2, filter_size, padding="same", activation="relu"),
+            Conv2D(n_filters*2, filter_size, padding="same", activation="relu"),
+            MaxPooling2D(pool_size=pool_size),
+            Dropout(conv_dropout),
 
-    train_shuffle_buffer_size = len(train_x)
-    validation_shuffle_buffer_size = len(validate_x)
+            Conv2D(n_filters*4, filter_size, padding="same", activation="relu"),
+            Conv2D(n_filters*4, filter_size, padding="same", activation="relu"),
+            MaxPooling2D(pool_size=pool_size),
+            Dropout(conv_dropout),
 
-    # Convert all y labels to numbers
-    train_processed_y = [label2index[label] for label in train_y]
-    validate_processed_y = [label2index[label] for label in validate_y]
-    test_processed_y = [label2index[label] for label in test_y]
+            Conv2D(n_filters*8, filter_size, padding="same", activation="relu"),
+            Conv2D(n_filters*8, filter_size, padding="same", activation="relu"),
+            MaxPooling2D(pool_size=pool_size),
+            Dropout(conv_dropout),
 
-    # Converts to y to binary class matrix (One-hot-encoded)
-    train_processed_y = to_categorical(
-        train_processed_y, num_classes=num_classes, dtype="float32"
+            Conv2D(n_filters*8, filter_size, padding="same", activation="relu"),
+            Conv2D(n_filters*8, filter_size, padding="same", activation="relu"),
+            MaxPooling2D(pool_size=pool_size),
+            Dropout(conv_dropout),
+
+            Flatten(),
+            Dense(dense_nodes, activation="relu"),
+            Dropout(dense_dropout),
+            Dense(n_classes, activation=output_activation),            
+        ],
+        name=f'{model_name}_{uuid.uuid4()}'
     )
-    validate_processed_y = to_categorical(
-        validate_processed_y, num_classes=num_classes, dtype="float32"
-    )
-    test_processed_y = to_categorical(
-        test_processed_y, num_classes=num_classes, dtype="float32"
-    )
+    
+    return model
 
-    # Create TF Dataset
-    train_data = tf.data.Dataset.from_tensor_slices((train_x, train_processed_y))
-    validation_data = tf.data.Dataset.from_tensor_slices(
-        (validate_x, validate_processed_y)
-    )
-    test_data = tf.data.Dataset.from_tensor_slices((test_x, test_processed_y))
 
-    #############
+
+# Parse a single image and label
+def parse_tfrecord_example(example):
+    # Read TF Records
+    feature_description = {
+        'image': tf.io.FixedLenFeature([], tf.string),
+        'label': tf.io.FixedLenFeature([5], tf.float32, default_value=[0.0, 0.0, 0.0, 0.0, 0.0])
+    }
+    
+    parsed_example = tf.io.parse_single_example(example, feature_description)
+
+    # Access global variables
+    global image_size, image_shape, num_classes
+
+    # Image
+    image = tf.io.decode_raw(parsed_example['image'], tf.uint8)
+    image.set_shape(image_size)
+    image = tf.reshape(image, image_shape)
+    
+    # Label
+    label = parsed_example['label']
+
+    return image, label
+
+
+
+# Normalize pixels
+def normalize(image, label):
+    image = image/255
+    return image, label
+
+
+
+# Load train and validation data from tfrecords files
+def load_tfrecords(
+        tfrecords_directory,
+        batch_size
+    ):
+
+    # Read the tfrecord files
+    train_tfrecord_files = tf.data.Dataset.list_files(tfrecords_directory+'/train*')
+    validate_tfrecord_files = tf.data.Dataset.list_files(tfrecords_directory+'/val*')
+
     # Train data
-    #############
-    # Apply all data processing logic
-    train_data = train_data.shuffle(buffer_size=train_shuffle_buffer_size)
-    train_data = train_data.map(load_image, num_parallel_calls=tf.data.AUTOTUNE)
+    train_data = train_tfrecord_files.flat_map(tf.data.TFRecordDataset)
+    train_data = train_data.map(parse_tfrecord_example, num_parallel_calls=tf.data.AUTOTUNE)
     train_data = train_data.map(normalize, num_parallel_calls=tf.data.AUTOTUNE)
     train_data = train_data.batch(batch_size)
-    train_data = train_data.prefetch(tf.data.AUTOTUNE)
+    train_data = train_data.prefetch(buffer_size=tf.data.AUTOTUNE)
 
-    ##################
     # Validation data
-    ##################
-    # Apply all data processing logic
-    validation_data = validation_data.shuffle(
-        buffer_size=validation_shuffle_buffer_size
-    )
-    validation_data = validation_data.map(
-        load_image, num_parallel_calls=tf.data.AUTOTUNE
-    )
-    validation_data = validation_data.map(
-        normalize, num_parallel_calls=tf.data.AUTOTUNE
-    )
+    validation_data = validate_tfrecord_files.flat_map(tf.data.TFRecordDataset)
+    validation_data = validation_data.map(parse_tfrecord_example, num_parallel_calls=tf.data.AUTOTUNE)
+    validation_data = validation_data.map(normalize, num_parallel_calls=tf.data.AUTOTUNE)
     validation_data = validation_data.batch(batch_size)
-    validation_data = validation_data.prefetch(tf.data.AUTOTUNE)
+    validation_data = validation_data.prefetch(buffer_size=tf.data.AUTOTUNE)
 
-    ############
-    # Test data
-    ############
-    # Apply all data processing logic
-    test_data = test_data.map(load_image, num_parallel_calls=tf.data.AUTOTUNE)
-    test_data = test_data.map(normalize, num_parallel_calls=tf.data.AUTOTUNE)
-    test_data = test_data.batch(batch_size)
-    test_data = test_data.prefetch(tf.data.AUTOTUNE)
+    print("train_data",train_data)
+    print("validation_data",validation_data)
 
-    return (train_data, validation_data, test_data)
+    return train_data, validation_data
 
 
-def build_mobilenet_model(
-    image_height, image_width, num_channels, num_classes, model_name, train_base=False
-):
-    # Model input
-    input_shape = [image_height, image_width, num_channels]  # height, width, channels
 
-    # Load a pretrained model from keras.applications
-    tranfer_model_base = keras.applications.MobileNetV2(
-        input_shape=input_shape, weights="imagenet", include_top=False
+def upload_model(model):
+    print(f'Uploading model to {GCS_BUCKET_NAME}')
+
+    # Create upload folder
+    upload_source = 'uploads'
+    os.makedirs(upload_source, exist_ok=True)
+
+    # Save model
+    file_name = f'{model.name}.h5'
+    model.save(os.path.join(upload_source, file_name))
+
+    # Instantiate GCS client
+    storage_client = storage.Client(project=GCP_PROJECT)
+    bucket = storage_client.bucket(GCS_BUCKET_NAME)
+
+    # Upload model
+    blob = bucket.blob(os.path.join(models_folder, file_name))
+    generation_match_precondition = 0
+    blob.upload_from_filename(
+        os.path.join(upload_source, file_name), 
+        if_generation_match=generation_match_precondition
+    )
+    print('Upload complete')    
+
+
+
+def download_tfrecords():
+    print(f'Downloading tfrecords files from {GCS_BUCKET_NAME}')
+
+    # Create download folder
+    download_destination = 'downloads'
+    os.makedirs(download_destination, exist_ok=True)
+
+    # Instantiate GCS client
+    client = storage.Client(project=GCP_PROJECT)
+
+    # Download tfrecords locally
+    blobs = client.list_blobs(GCS_BUCKET_NAME, prefix=tfrecords_folder)
+    for blob in blobs:
+        if blob.name.endswith("/"):
+            continue
+        print(f'Downloading: {blob.name}')
+        blob.download_to_filename(
+            os.path.join(download_destination, os.path.basename(blob.name))
+        )
+    print('Download complete')
+
+
+# Read config file
+with open(args.config, 'r') as file:
+    config = yaml.full_load(file)
+
+
+# Build model
+model = build_model(**config['build_params'])
+print(model.summary())
+
+
+# Set optimizer
+if config['compile_params']['optimizer'] == 'adam':
+    optimizer = Adam(
+        learning_rate = config['compile_params']['learning_rate']
     )
 
-    # Freeze the mobileNet model layers
-    tranfer_model_base.trainable = train_base
 
-    # Regularize using L1
-    kernel_weight = 0.02
-    bias_weight = 0.02
-
-    model = Sequential(
-        [
-            tranfer_model_base,
-            keras.layers.GlobalAveragePooling2D(),
-            keras.layers.Dense(
-                units=128,
-                activation="relu",
-                kernel_regularizer=keras.regularizers.l1(kernel_weight),
-                bias_regularizer=keras.regularizers.l1(bias_weight),
-            ),
-            keras.layers.Dense(
-                units=num_classes,
-                activation="softmax",
-                kernel_regularizer=keras.regularizers.l1(kernel_weight),
-                bias_regularizer=keras.regularizers.l1(bias_weight),
-            ),
-        ],
-        name=model_name + "_train_base_" + str(train_base),
-    )
-
-    return model
+# Set metrics
+for i, metric in enumerate(config['compile_params']['metrics']):
+    if metric ==  'rmse':
+        config['compile_params']['metrics'][i] = tf.keras.metrics.RootMeanSquaredError()
 
 
-def build_model_tfhub(
-    image_height, image_width, num_channels, num_classes, model_name, train_base=False
-):
-    # Model input
-    input_shape = [image_height, image_width, num_channels]  # height, width, channels
-
-    # Handle to pretrained model
-    handle = "https://tfhub.dev/google/imagenet/mobilenet_v2_100_224/feature_vector/4"
-
-    # Regularize using L1
-    kernel_weight = 0.02
-    bias_weight = 0.02
-
-    model = Sequential(
-        [
-            keras.layers.InputLayer(input_shape=input_shape),
-            hub.KerasLayer(handle, trainable=train_base),
-            keras.layers.Dense(
-                units=64,
-                activation="relu",
-                kernel_regularizer=keras.regularizers.l1(kernel_weight),
-                bias_regularizer=keras.regularizers.l1(bias_weight),
-            ),
-            keras.layers.Dense(
-                units=num_classes,
-                activation="softmax",
-                kernel_regularizer=keras.regularizers.l1(kernel_weight),
-                bias_regularizer=keras.regularizers.l1(bias_weight),
-            ),
-        ],
-        name=model_name + "_train_base_" + str(train_base),
-    )
-
-    return model
-
-
-print("Train model")
-############################
-# Training Params
-############################
-model_name = args.model_name
-learning_rate = 0.001
-image_width = 224
-image_height = 224
-num_channels = 3
-batch_size = args.batch_size
-epochs = args.epochs
-train_base = args.train_base
-
-# Free up memory
-K.clear_session()
-
-# Data
-train_data, validation_data, test_data = get_dataset(
-    image_width=image_width,
-    image_height=image_height,
-    num_channels=num_channels,
-    batch_size=batch_size,
+# Compile model
+model.compile(
+    loss = config['compile_params']['loss'],
+    optimizer = optimizer,
+    metrics = config['compile_params']['metrics']
 )
 
-if model_name == "mobilenetv2":
-    # Model
-    model = build_mobilenet_model(
-        image_height,
-        image_width,
-        num_channels,
-        num_classes,
-        model_name,
-        train_base=train_base,
-    )
-    # Optimizer
-    optimizer = keras.optimizers.SGD(learning_rate=learning_rate)
-    # Loss
-    loss = keras.losses.categorical_crossentropy
-    # Print the model architecture
-    print(model.summary())
-    # Compile
-    model.compile(loss=loss, optimizer=optimizer, metrics=["accuracy"])
-elif model_name == "tfhub_mobilenetv2":
-    # Model
-    model = build_model_tfhub(
-        image_height,
-        image_width,
-        num_channels,
-        num_classes,
-        model_name,
-        train_base=train_base,
-    )
-    # Optimizer
-    optimizer = keras.optimizers.SGD(learning_rate=learning_rate)
-    # Loss
-    loss = keras.losses.categorical_crossentropy
-    # Print the model architecture
-    print(model.summary())
-    # Compile
-    model.compile(loss=loss, optimizer=optimizer, metrics=["accuracy"])
 
-# Initialize a W&B run
+# Define global variables for tfrecords parser
+global image_size, image_shape, num_classes
+image_size = [math.prod(config['build_params']['input_shape'])]
+image_shape = list(config['build_params']['input_shape'])
+num_classes = config['build_params']['n_classes']
+
+
+# Fetch data
+download_tfrecords()
+train_data, validation_data = load_tfrecords(
+    'downloads',
+    config['train_params']['batch_size']
+)
+
+
+# Initialize a Weights and Biases run
 wandb.init(
-    project="mushroom-training-vertex-ai",
+    project="snapnutrition-training-vertex-ai",
     config={
-        "learning_rate": learning_rate,
-        "epochs": epochs,
-        "batch_size": batch_size,
+        "learning_rate": config['compile_params']['learning_rate'],
+        "epochs": config['train_params']['epochs'],
+        "batch_size": config['train_params']['batch_size'],
         "model_name": model.name,
     },
     name=model.name,
@@ -385,17 +298,24 @@ wandb.init(
 # Train model
 start_time = time.time()
 training_results = model.fit(
-    train_data,
-    validation_data=validation_data,
-    epochs=epochs,
-    callbacks=[WandbCallback()],
-    verbose=1,
+        train_data,
+        validation_data = validation_data,
+        epochs = config['train_params']['epochs'],
+        callbacks=[WandbCallback()],
+        verbose = 1
 )
-execution_time = (time.time() - start_time) / 60.0
-print("Training execution time (mins)", execution_time)
+execution_time = (time.time() - start_time)/60.0
+print(f'Training execution time (mins): {execution_time:.02f}')
+
+
+# Upload model
+upload_model(model)
+
 
 # Update W&B
 wandb.config.update({"execution_time": execution_time})
+
+
 # Close the W&B run
 wandb.run.finish()
